+++ braintree/payment_instrument_type.py
@@ -1,17 +1,18 @@
 
 class PaymentInstrumentType():
+    # NEXT_MAJOR_VERSION remove amex express checkout
+    # NEXT_MAJOR_VERSION remove masterpass
+    AmexExpressCheckoutCard = "amex_express_checkout_card"
+    AndroidPayCard = "android_pay_card"
+    ApplePayCard = "apple_pay_card"
+    CreditCard = "credit_card"
+    EuropeBankAccount = "europe_bank_account"
+    LocalPayment = "local_payment"
+    MasterpassCard = "masterpass_card"
     PayPalAccount = "paypal_account"
     PayPalHere = "paypal_here"
-    EuropeBankAccount = "europe_bank_account"
-    CreditCard = "credit_card"
-    ApplePayCard = "apple_pay_card"
-    AndroidPayCard = "android_pay_card"
-    # NEXT_MAJOR_VERSION remove amex express checkout
-    AmexExpressCheckoutCard = "amex_express_checkout_card"
-    VenmoAccount = "venmo_account"
-    UsBankAccount = "us_bank_account"
-    VisaCheckoutCard = "visa_checkout_card"
-    # NEXT_MAJOR_VERSION remove masterpass
-    MasterpassCard = "masterpass_card"
     SamsungPayCard = "samsung_pay_card"
-    LocalPayment = "local_payment"
+    SepaDirectDebitAccount = "sepa_debit_account"
+    UsBankAccount = "us_bank_account"
+    VenmoAccount = "venmo_account"
+    VisaCheckoutCard = "visa_checkout_card"